# openapi.json
{"openapi":"3.1.0","info":{"title":"Portfolio API","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/holdings":{"get":{"summary":"List Holdings","operationId":"list_holdings_holdings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/HoldingOut"},"type":"array","title":"Response List Holdings Holdings Get"}}}}}},"post":{"summary":"Create Holding","operationId":"create_holding_holdings_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HoldingIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HoldingOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/holdings/{holding_id}":{"delete":{"summary":"Delete Holding","operationId":"delete_holding_holdings__holding_id__delete","parameters":[{"name":"holding_id","in":"path","required":true,"schema":{"type":"string","title":"Holding Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Delete Holding Holdings  Holding Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Holding","operationId":"get_holding_holdings__holding_id__get","parameters":[{"name":"holding_id","in":"path","required":true,"schema":{"type":"string","title":"Holding Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HoldingOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Holding","operationId":"update_holding_holdings__holding_id__put","parameters":[{"name":"holding_id","in":"path","required":true,"schema":{"type":"string","title":"Holding Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HoldingIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HoldingOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/quote":{"get":{"summary":"Get Quote","operationId":"get_quote_quote_get","parameters":[{"name":"symbol","in":"query","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search/{symbol}":{"get":{"summary":"Search Symbol","operationId":"search_symbol_search__symbol__get","parameters":[{"name":"symbol","in":"path","required":true,"schema":{"type":"string","title":"Symbol"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HoldingIn":{"properties":{"symbol":{"type":"string","title":"Symbol"},"shares":{"type":"number","title":"Shares"},"avgCost":{"type":"number","title":"Avgcost"},"currency":{"type":"string","title":"Currency","default":"USD"}},"type":"object","required":["symbol","shares","avgCost"],"title":"HoldingIn"},"HoldingOut":{"properties":{"symbol":{"type":"string","title":"Symbol"},"shares":{"type":"number","title":"Shares"},"avgCost":{"type":"number","title":"Avgcost"},"currency":{"type":"string","title":"Currency","default":"USD"},"id":{"type":"string","title":"Id"}},"type":"object","required":["symbol","shares","avgCost","id"],"title":"HoldingOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}